Moved Dto's to a client package
